fix: changes as per comments
(cherry picked from commit 42d9beed1a0ae90e48a3cf66b34294a1c37f8fdf)
