+++ india_compliance/gst_india/client_scripts/company.js
@@ -7,6 +7,7 @@ validate_gstin(DOCTYPE);
 update_gstin_in_other_documents(DOCTYPE);
 show_overseas_disabled_warning(DOCTYPE);
 set_gstin_options_and_status(DOCTYPE);
+set_pan_status(DOCTYPE)
 
 frappe.ui.form.off(DOCTYPE, "make_default_tax_template");
 frappe.ui.form.on(DOCTYPE, {
