+++ india_compliance/gst_india/client_scripts/supplier.js
@@ -8,6 +8,7 @@ update_gstin_in_other_documents(DOCTYPE);
 show_overseas_disabled_warning(DOCTYPE);
 set_gstin_options_and_status(DOCTYPE);
 set_gst_category(DOCTYPE);
+set_pan_status(DOCTYPE)
 
 frappe.ui.form.on(DOCTYPE, {
     gstin(frm) {
